[refactor]: rename operator module class file names snake_case
